+++ npdependency/deptree.py
@@ -194,3 +194,181 @@ class DepGraph:
 
     def __len__(self):
         return len(self.words)
+
+
+class DependencyDataset:
+    """
+    A representation of the DepBank for efficient processing.
+    This is a sorted dataset.
+    """
+    PAD_IDX            = 0
+    PAD_TOKEN          = '<pad>'
+    UNK_WORD           = '<unk>'
+
+    @staticmethod
+    def read_conll(filename):
+        istream       = open(filename)
+        treelist = []
+        tree = DepGraph.read_tree(istream) 
+        while tree:
+            if len(tree.words) <= 150:
+                treelist.append(tree)
+            else:
+                print('dropped tree with length',len(tree.words))
+            tree = DepGraph.read_tree(istream)
+        istream.close()  
+        return treelist
+        
+    def __init__(self,treelist,lexer,char_dataset,ft_dataset,use_labels=None,use_tags=None):
+        self.lexer        = lexer
+        self.char_dataset = char_dataset
+        self.ft_dataset   = ft_dataset
+        self.treelist     = treelist
+        if use_labels:
+            self.itolab = use_labels
+            self.labtoi = {label:idx for idx,label in enumerate(self.itolab)}
+        else:
+            self.init_labels(self.treelist)
+        if use_tags:
+            self.itotag = use_tags
+            self.tagtoi = {tag:idx for idx,tag in enumerate(self.itotag)}
+        else:
+            self.init_tags(self.treelist)
+
+    def encode(self):
+        self.deps, self.heads,self.labels,self.tags = [ ],[ ],[ ],[ ]
+        self.words,self.mwe_ranges, self.cats = [ ], [ ], [ ]
+
+        for tree in self.treelist:
+            depword_idxes = self.lexer.tokenize(tree.words)
+            deptag_idxes  = [self.tagtoi.get(tag,self.tagtoi[DependencyDataset.UNK_WORD]) for tag in tree.pos_tags]
+
+            self.words.append(tree.words)
+            self.cats.append(tree.pos_tags)
+            self.tags.append(deptag_idxes)
+            self.deps.append(depword_idxes)
+            self.heads.append(self.oracle_governors(tree))
+            # the get defaulting to 0 is a hack for labels not found in training set
+            self.labels.append([self.labtoi.get(lab,0) for lab in self.oracle_labels(tree)])
+            self.mwe_ranges.append(tree.mwe_ranges)
+
+    def save_vocab(self,filename):
+        out = open(filename,'w')
+        print(' '.join(self.itolab),file=out)
+        print(' '.join(self.itotag),file=out)
+        out.close()
+
+    @staticmethod
+    def load_vocab(filename):
+        reloaded = open(filename)
+        itolab   = reloaded.readline().split()
+        itotag   = reloaded.readline().split()
+        reloaded.close()
+        return itos,itolab,itotag
+    
+    def shuffle_data(self):
+        N = len(self.deps)
+        order = list(range(N))
+        shuffle(order)
+        self.deps       = [self.deps[i] for i in order]
+        self.tags       = [self.tags[i] for i in order]
+        self.heads      = [self.heads[i] for i in order]
+        self.labels     = [self.labels[i] for i in order]
+        self.words      = [self.words[i] for i in order]
+        self.cats       = [self.cats[i] for i in order]
+        self.mwe_ranges = [self.mwe_ranges[i] for i in order]
+
+    def order_data(self):
+        N           = len(self.deps)
+        order       = list(range(N))        
+        lengths     = map(len,self.deps)
+        order       = [idx for idx, L in sorted(zip(order,lengths),key=lambda x:x[1])]
+        self.deps   = [self.deps[idx] for idx in order]
+        self.tags   = [self.tags[idx] for idx in order]
+        self.heads  = [self.heads[idx] for idx in order]
+        self.labels = [self.labels[idx] for idx in order]
+        self.words  = [self.words[idx] for idx in order]
+        self.mwe_ranges  = [self.mwe_ranges[idx] for idx in order]
+        self.cats        = [self.cats[idx] for idx in order]
+        
+    def make_batches(self, batch_size,shuffle_batches=False,shuffle_data=True,order_by_length=False):
+        self.encode()
+        if shuffle_data:  
+            self.shuffle_data()
+        if order_by_length: #shuffling and ordering is relevant : it change the way ties are resolved and thus batch construction
+            self.order_data()
+
+        N = len(self.deps)
+        batch_order = list(range(0,N, batch_size))
+        if shuffle_batches:
+            shuffle(batch_order)
+        for i in batch_order:
+            deps     = self.pad(self.deps[i:i+batch_size])
+            tags     = self.pad(self.tags[i:i+batch_size])
+            heads    = self.pad(self.heads[i:i+batch_size])
+            labels   = self.pad(self.labels[i:i+batch_size])
+            words    = self.words[i:i+batch_size]
+            mwe      = self.mwe_ranges[i:i+batch_size]
+            cats     = self.cats[i:i+batch_size]
+            chars    = self.char_dataset.batch_chars(self.words[i:i+batch_size])
+            subwords = self.ft_dataset.batch_sentences(self.words[i:i+batch_size])
+            yield (words,mwe,chars,subwords,cats,deps,tags,heads,labels)
+
+    def pad(self,batch): 
+        if type(batch[0]) == tuple and len(batch[0]) == 2:   #had hoc stuff for BERT Lexers
+            sent_lengths                = [ len(seqA) for (seqA,seqB) in batch] 
+            max_len                     = max(sent_lengths)
+            padded_batchA,padded_batchB = [ ], [ ]
+            for k, seq in zip(sent_lengths, batch):
+                seqA,seqB = seq
+                paddedA   = seqA + (max_len - k)*[ DependencyDataset.PAD_IDX ]
+                paddedB   = seqB + (max_len - k)*[ self.lexer.BERT_PAD_IDX ]  
+                padded_batchA.append(paddedA)
+                padded_batchB.append(paddedB)
+            return  ( Variable(torch.LongTensor(padded_batchA)) , Variable(torch.LongTensor(padded_batchB)))
+        else:
+            sent_lengths = list(map(len, batch))
+            max_len      = max(sent_lengths)
+            padded_batch = [ ]
+            for k, seq in zip(sent_lengths, batch):
+                padded = seq + (max_len - k) * [ DependencyDataset.PAD_IDX]
+                padded_batch.append(padded)
+        return Variable( torch.LongTensor(padded_batch) )
+
+    def init_labels(self,treelist):
+        labels      = set([ lbl for tree in treelist for (gov,lbl,dep) in tree.get_all_edges()])
+        self.itolab = [DependencyDataset.PAD_TOKEN] + list(labels)
+        self.labtoi = {label:idx for idx,label in enumerate(self.itolab)}
+
+    def init_tags(self,treelist):
+        tagset  = set([ tag for tree in treelist for tag in tree.pos_tags])
+        tagset.update([DepGraph.ROOT_TOKEN,DependencyDataset.UNK_WORD])
+        self.itotag = [DependencyDataset.PAD_TOKEN] + list(tagset)
+        self.tagtoi = {tag:idx for idx,tag in enumerate(self.itotag)}
+            
+    def __len__(self):      
+        return len(self.treelist)
+    
+    def oracle_labels(self,depgraph):
+        """
+        Returns a list where each element list[i] is the label of
+        the position of the governor of the word at position i.
+        Returns:
+        a tensor of size N.
+        """
+        N          = len(depgraph)
+        edges      = depgraph.get_all_edges( )
+        rev_labels = dict( [ (dep,label) for (gov,label,dep) in edges ] )
+        return [ rev_labels.get(idx,DependencyDataset.PAD_TOKEN) for idx in range(N) ]   
+        
+    def oracle_governors( self,depgraph ): 
+        """
+        Returns a list where each element list[i] is the index of
+        the position of the governor of the word at position i.
+        Returns:
+        a tensor of size N.
+        """
+        N         = len( depgraph )
+        edges     = depgraph.get_all_edges( )
+        rev_edges = dict( [(dep,gov) for (gov,label,dep) in edges ] )
+        return [ rev_edges.get(idx,0) for idx in range(N) ]
